Update IntuneMunkiManifestGenerator.py
Version 1.3: 2021-12-06 - Added check for null AAD device ID, removal of department manifest if device is no longer in group and
removal of manifest if device is no longer in Intune. If you want to do this, un-comment line 300. Run one time first and check output 
to make sure manifests that should not be deleted are touched.